+++ misc/math_utils.py
@@ -24,6 +24,8 @@ import scipy.stats
 import sklearn
 import sklearn.model_selection
 
+import misc.validation_utils as validation_utils
+
 import logging
 logger = logging.getLogger(__name__)
 
@@ -937,7 +939,84 @@ def get_categorical_mle_estimates(
         
     mle_estimates = mle_estimates / np.sum(mle_estimates)
     return mle_estimates
+   
+def collect_regression_metrics(y_true, y_pred):
+    """ Collect various classification performance metrics for the predictions
+
+    Parameters
+    ----------
+    y_true: np.array of real values
+        The true value of each instance
+
+    y_pred: np.array of floats
+        The prediction for each instance
     
+    Returns
+    -------
+    metrics: dict
+        A mapping from the metric name to the respective value
+    """
+    validation_utils.validate_equal_shape(y_true, y_pred)
+
+    ret = {
+        "explained_variance": sklearn.metrics.explained_variance_score(y_true, y_pred),
+        "mean_absolute_error": sklearn.metrics.mean_absolute_error(y_true, y_pred),
+        "mean_squared_error": sklearn.metrics.mean_squared_error(y_true, y_pred),
+        #"mean_squared_log_error": sklearn.metrics.mean_squared_log_error(y_true, y_pred),
+        "median_absolute_error": sklearn.metrics.median_absolute_error(y_true, y_pred),
+        "r2": sklearn.metrics.r2_score(y_true, y_pred)
+    }
+
+    return ret
+
+
+def collect_multiclass_classification_metrics(y_true, y_score):
+    """ Calculate various multi-class classification performance metrics
+    
+    Parameters
+    ----------
+    y_true: np.array with shape [n_samples]
+        The true label of each instance. The labels are assumed to be encoded 
+        with integers [0, 1, ... n_classes-1]. The respective columns in
+        y_score should give the scores of the matching label.
+        
+    y_score: np.array with shape [n_samples, n_classes]
+        The score predictions for each class, e.g., from pred_proba, though
+        they are not required to be probabilities
+        
+    Returns
+    -------
+    metrics: dict
+        A mapping from the metric name to the respective value
+    """
+
+    # make hard predictions
+    y_pred = np.argmax(y_score, axis=1)
+
+    # now collect all statistics
+    ret = {
+         "cohen_kappa":  sklearn.metrics.cohen_kappa_score(y_true, y_pred),
+         "matthews_corrcoef":  sklearn.metrics.matthews_corrcoef(y_true, y_pred),
+         "accuracy":  sklearn.metrics.accuracy_score(y_true, y_pred),
+         "micro_f1_score":  sklearn.metrics.f1_score(y_true, y_pred,
+            average='micro'),
+         "macro_f1_score":  sklearn.metrics.f1_score(y_true, y_pred,
+            average='macro'),
+         "hamming_loss":  sklearn.metrics.hamming_loss(y_true, y_pred),
+         "micro_precision":  sklearn.metrics.precision_score(y_true, y_pred,
+            average='micro'),
+         "macro_precision":  sklearn.metrics.precision_score(y_true, y_pred,
+            average='macro'),
+         "micro_recall":  sklearn.metrics.recall_score(y_true, y_pred,
+            average='micro'),
+         "macro_recall":  sklearn.metrics.recall_score(y_true, y_pred,
+            average='macro'),
+         "hand_and_till_m_score": calc_hand_and_till_m_score(y_true, y_score),
+         "provost_and_domingos_auc": calc_provost_and_domingos_auc(y_true, y_score)
+    }
+
+    return ret
+
 
 def collect_binary_classification_metrics(y_true, y_probas_pred, threshold=0.5,
         pos_label=1):
@@ -1266,14 +1345,16 @@ fold_tuple_fields = [
 ]
 fold_tuple = collections.namedtuple('fold', ' '.join(fold_tuple_fields))
 
-def get_kth_fold(X, y, fold, num_folds=10, random_seed=8675309):
-    """ Select the kth cross-validation fold using stratified CV
+def get_kth_fold(X, y, fold, num_folds=10, use_stratified=True,
+        random_seed=8675309):
+    """ Select the kth cross-validation fold using (stratified) CV
     
     In partcular, this function uses `sklearn.model_selection.StratifiedKFold`
     to split the data. It then selects the training and testing splits
     from the k^th fold.
 
-    N.B. If `y` is None, the simple `KFold` is used instead.
+    N.B. If `y` is None, or `use_stratified` is False, then simple `KFold` is
+    used instead.
     
     Parameters
     ----------
@@ -1284,6 +1365,10 @@ def get_kth_fold(X, y, fold, num_folds=10, random_seed=8675309):
         
     num_folds: int
         The total number of folds
+
+    use_stratified: bool
+        Whether to use stratified cross-validation. For example, this may be
+        set to False if choosing folds for regression.
         
     random_seed: int or random state
         The value used a the random seed for the k-fold split
@@ -1291,7 +1376,7 @@ def get_kth_fold(X, y, fold, num_folds=10, random_seed=8675309):
 
     check_range(fold, 0, num_folds, max_inclusive=False, variable_name='fold')
 
-    if y is None:
+    if (y is None) or (not use_stratified):
         cv = sklearn.model_selection.KFold(
             n_splits=num_folds, random_state=random_seed
         )