+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Helper to check if class attributes have been initialized similar to
     `check_is_fitted` in sklearn
 - Standard validation helpers (`validation_utils`)
+- Helper to calculate many regression, multi-class classification metrics
 
 ## [0.2.8] - 2018-03-12
 ### Updated